Added some checks to the competency challenge problem.

+++ Pathways/Stage 2/Week-3/CompChalProbDIandTesting/RecipesCollection/IPayForMeal.cs
@@ -8,6 +8,6 @@ namespace Meals
 {
     public interface IPayForMeal
     {
-        int PayForMeal(int items);
+        double PayForMeal(double items);
     }
 }
